Modify SubBlock class to combine in/out pins...
Modify SubBlock class to combine input and output pins into a single vector.
This is necessary since VPR v4.3 assumes a single subblock output pin, whereas
VPR v5 allows for a variable number of input and output subblock pins.  This
way, the division of the pins is left up to the client program instead of being
directly handled by the parser class.